m5: phase brackets for the IPC round trip, Order as a bitmask
- expose the bench helpers through a KOS_SYS_BENCH syscall; bench.h holds the
  per-arch cycle source and the phase table the kernel prints
- bracket endpoint_call and endpoint_reply per phase, and run one bench step
  with the caller above the server so the donate branch is reached
- priority donation costs 149 cycles on the M7 and 160 on the lx6. The fastpath
  never reaches it: a caller outranking its server takes the slowpath on every call
- make Order a bitmask; head and tail carry their ordering in their type and
  the byte ring's publication-barrier macro goes
- narrow Thread::switch_count into the padding hole before deadline_ns,
  sizeof(Thread) 264 to 256 bytes on armv7m
- add tests/unit/capprobe, gating the reply mint's probe against the install it stands for

+++ kernel/include/kickos/sync.h
@@ -68,7 +68,7 @@ namespace kickos
 
     // Resume barrier, MANDATORY for any blocking primitive that reads waker-set TCB state
     // (wait_result) after resuming:
-    //     Thread* c = sched::current(); uint64_t epoch;
+    //     Thread* c = sched::current(); uint32_t epoch;
     //     { IrqLock lock; ...predicate + set up state...; epoch = c->switch_count;
     //       wq_block(q, kind, obj); }              // lock RELEASED here
     //     wq_confirm_resume(c, epoch);              // <- barrier, OUTSIDE the lock
@@ -78,7 +78,7 @@ namespace kickos
     // drops, and a wait_result read there returns the PRE-block value. No-op on the sim,
     // whose switch is synchronous. The waker, never the sleeper, must write wait_result
     // and clear the wait edge under the lock.
-    void wq_confirm_resume(Thread* c, uint64_t epoch);
+    void wq_confirm_resume(Thread* c, uint32_t epoch);
 
     void sem_init(Semaphore* s, int initial);
     void sem_wait(Semaphore* s);
